requirements.txt
python-telegram-bot==21.3
httpx==0.28.1
groq==1.2.0
google-api-python-client==2.194.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
notion-client==2.2.1
todoist-api-python==4.0.0
python-dotenv==1.2.2
aiofiles==25.1.0